requirements.txt
aioredis==2.0.1
fastapi==0.115.12
odmantic==1.0.2
orjson==3.10.16
starlette==0.46.1
uvicorn==0.19.0
pydantic==2.11.1
pyarmor==8.4.7
dill==0.3.7
durable_rules==2.0.28
httpx==0.28.1
langchain==0.3.20
langchain_community==0.3.19
langchain_openai==0.3.7
networkx==3.2.1
numpy==1.26.2
openai==1.65.4
pandas==2.0.3
py2neo==2021.2.4
pypinyin==0.53.0
python-dotenv==1.0.1
pyvis==0.3.2
scipy==1.11.1
scikit_learn==1.3.2
faiss-cpu==1.10.0